clean du code et correction de plusieurs bugs

+++ ServeurWeb/app/src/main/java/com/analyse_crypto/app/tables/utilisateurs/repository/PortefeuilleRepository.java
@@ -9,5 +9,5 @@ import com.analyse_crypto.app.tables.utilisateurs.User;
 
 public interface PortefeuilleRepository extends JpaRepository<Portefeuille, Long> {
     List<Portefeuille> findByUser(User user);
-    Portefeuille findByIdPortefeuilleAndUser(int id, User user);
+    Portefeuille findByIdPortefeuilleAndUser(Long id, User user);
 }
